+++ docs/source/CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- `toff.utils.safe_naming`. Add a prefix to the atom types.
+- `safe_naming_prefix` keyword on `toff.utils.Parameterize`. It uses the previous function if some string is provided.
+
 ### Fixed
 
 - Improve documentation
-
 
 [unreleased]: https://github.com/ale94mleon/TOFF/compare/0.0.0-alpha2...HEAD